AzureBlobStorage/requirements.txt
aiohttp==3.6.2
azure-storage-blob==12.3.1
python-dotenv==0.12.0